Genesis 3: mark wide ranging constraints with wide flag
Mark wide ranging constraints in Genesis 3 with the flags([wide])
source, specifically the fact that Eve was the mother of all living, and
that everybody was conceived after the closing of Eden.

This is used when generating diagrams to avoid spaghettification due to
lines from a single node to almost every other node.

Signed-off-by: James Hogan <james@albanarts.com>